--- src/mock/destinations-mock.js
@@ -1,150 +1,0 @@
-/* eslint-disable quotes */
-export const destinationsMocks = [
-  {
-    "id": "b4f450d6-44d5-420d-be1e-a24bbf4516a1",
-    "description": "Vien - middle-eastern paradise",
-    "name": "Vien",
-    "pictures": [
-      {
-        "src": "https://24.objects.htmlacademy.pro/static/destinations/16.jpg",
-        "description": "Vien full of of cozy canteens where you can try the best coffee in the Middle East"
-      },
-      {
-        "src": "https://24.objects.htmlacademy.pro/static/destinations/14.jpg",
-        "description": "Vien is a beautiful city"
-      }
-    ]
-  },
-  {
-    "id": "f938701f-6c0c-4404-b46c-60134a8cf73f",
-    "description": "Nagasaki - with crowded streets",
-    "name": "Nagasaki",
-    "pictures": []
-  },
-  {
-    "id": "7f7f2401-27d7-4969-9736-7c84f65337e5",
-    "description": "Kopenhagen - full of of cozy canteens where you can try the best coffee in the Middle East",
-    "name": "Kopenhagen",
-    "pictures": [
-      {
-        "src": "https://24.objects.htmlacademy.pro/static/destinations/15.jpg",
-        "description": "Kopenhagen famous for its crowded street markets with the best street food in Asia"
-      },
-      {
-        "src": "https://24.objects.htmlacademy.pro/static/destinations/16.jpg",
-        "description": "Kopenhagen with an embankment of a mighty river as a centre of attraction"
-      },
-      {
-        "src": "https://24.objects.htmlacademy.pro/static/destinations/5.jpg",
-        "description": "Kopenhagen a true asian pearl"
-      },
-      {
-        "src": "https://24.objects.htmlacademy.pro/static/destinations/7.jpg",
-        "description": "Kopenhagen famous for its crowded street markets with the best street food in Asia"
-      }
-    ]
-  },
-  {
-    "id": "32a119ce-7ec8-44e3-a7dc-e16c0e63ff25",
-    "description": "Rotterdam - full of of cozy canteens where you can try the best coffee in the Middle East",
-    "name": "Rotterdam",
-    "pictures": []
-  },
-  {
-    "id": "b641fae0-6a67-42ac-910e-ea22b152d62e",
-    "description": "Tokio - with crowded streets",
-    "name": "Tokio",
-    "pictures": [
-      {
-        "src": "https://24.objects.htmlacademy.pro/static/destinations/8.jpg",
-        "description": "Tokio with a beautiful old town"
-      },
-      {
-        "src": "https://24.objects.htmlacademy.pro/static/destinations/4.jpg",
-        "description": "Tokio in a middle of Europe"
-      }
-    ]
-  },
-  {
-    "id": "87047be5-3d94-42ee-8401-a1419dd45abf",
-    "description": "Barcelona - full of of cozy canteens where you can try the best coffee in the Middle East",
-    "name": "Barcelona",
-    "pictures": [
-      {
-        "src": "https://24.objects.htmlacademy.pro/static/destinations/17.jpg",
-        "description": "Barcelona in a middle of Europe"
-      },
-      {
-        "src": "https://24.objects.htmlacademy.pro/static/destinations/10.jpg",
-        "description": "Barcelona with a beautiful old town"
-      },
-      {
-        "src": "https://24.objects.htmlacademy.pro/static/destinations/13.jpg",
-        "description": "Barcelona with an embankment of a mighty river as a centre of attraction"
-      },
-      {
-        "src": "https://24.objects.htmlacademy.pro/static/destinations/19.jpg",
-        "description": "Barcelona a perfect place to stay with a family"
-      },
-      {
-        "src": "https://24.objects.htmlacademy.pro/static/destinations/1.jpg",
-        "description": "Barcelona famous for its crowded street markets with the best street food in Asia"
-      }
-    ]
-  },
-  {
-    "id": "ec1dd8de-dfd5-49cd-b0fd-bc6ddfb74f34",
-    "description": "Kioto - for those who value comfort and coziness",
-    "name": "Kioto",
-    "pictures": [
-      {
-        "src": "https://24.objects.htmlacademy.pro/static/destinations/3.jpg",
-        "description": "Kioto with an embankment of a mighty river as a centre of attraction"
-      },
-      {
-        "src": "https://24.objects.htmlacademy.pro/static/destinations/13.jpg",
-        "description": "Kioto is a beautiful city"
-      },
-      {
-        "src": "https://24.objects.htmlacademy.pro/static/destinations/18.jpg",
-        "description": "Kioto famous for its crowded street markets with the best street food in Asia"
-      },
-      {
-        "src": "https://24.objects.htmlacademy.pro/static/destinations/14.jpg",
-        "description": "Kioto for those who value comfort and coziness"
-      },
-      {
-        "src": "https://24.objects.htmlacademy.pro/static/destinations/11.jpg",
-        "description": "Kioto a true asian pearl"
-      }
-    ]
-  },
-  {
-    "id": "c4586880-c31c-45b2-8ef1-9cb807038ef0",
-    "description": "Helsinki - with crowded streets",
-    "name": "Helsinki",
-    "pictures": [
-      {
-        "src": "https://24.objects.htmlacademy.pro/static/destinations/18.jpg",
-        "description": "Helsinki with a beautiful old town"
-      }
-    ]
-  },
-  {
-    "id": "6b2fb21b-bbf2-40af-95b3-10f43728b69f",
-    "description": "Hiroshima - famous for its crowded street markets with the best street food in Asia",
-    "name": "Hiroshima",
-    "pictures": [
-      {
-        "src": "https://24.objects.htmlacademy.pro/static/destinations/17.jpg",
-        "description": "Hiroshima a perfect place to stay with a family"
-      }
-    ]
-  },
-  {
-    "id": "fd6dcceb-6625-428f-b0d8-31f4ecb2dcc4",
-    "description": "Saint Petersburg - a true asian pearl",
-    "name": "Saint Petersburg",
-    "pictures": []
-  }
-];
