--- src/mock/offers-mock.js
@@ -1,207 +1,0 @@
-/* eslint-disable quotes */
-export const offersMocks = [
-  {
-    "type": "taxi",
-    "offers": [
-      {
-        "id": "21e8fecf-2cf7-472a-846b-75d8787d64c9",
-        "title": "Upgrade to a business class",
-        "price": 51
-      },
-      {
-        "id": "77ca1839-c4ba-44e6-ab5f-85a37f097b1c",
-        "title": "Choose the radio station",
-        "price": 104
-      },
-      {
-        "id": "059efaf8-3db3-41b4-a4ad-1c6764f23604",
-        "title": "Choose temperature",
-        "price": 87
-      },
-      {
-        "id": "7a74643a-173b-4cc6-90ff-cf3e7cc7cef5",
-        "title": "Drive quickly, I'm in a hurry",
-        "price": 46
-      },
-      {
-        "id": "bfe7c32e-e1a8-4ff9-9629-921e3958369a",
-        "title": "Drive slowly",
-        "price": 196
-      }
-    ]
-  },
-  {
-    "type": "bus",
-    "offers": [
-      {
-        "id": "b1caa362-4fcb-46e9-8476-92929e9e505f",
-        "title": "Infotainment system",
-        "price": 138
-      },
-      {
-        "id": "515272d6-4a76-4248-9dcc-634f97b86111",
-        "title": "Order meal",
-        "price": 156
-      },
-      {
-        "id": "8c4b9cd9-fef2-44db-8d61-daf388ed8c2c",
-        "title": "Choose seats",
-        "price": 63
-      }
-    ]
-  },
-  {
-    "type": "train",
-    "offers": [
-      {
-        "id": "8267a6b8-1f08-4bc8-a298-d6eb34417ad3",
-        "title": "Book a taxi at the arrival point",
-        "price": 142
-      },
-      {
-        "id": "42e59c87-a3a5-43a6-8790-faebd2892b3b",
-        "title": "Order a breakfast",
-        "price": 135
-      },
-      {
-        "id": "0630a002-7dce-4ad1-be9e-7e8b64152a97",
-        "title": "Wake up at a certain time",
-        "price": 130
-      }
-    ]
-  },
-  {
-    "type": "flight",
-    "offers": [
-      {
-        "id": "16ca7457-5366-4b4c-b903-895a3adeaad2",
-        "title": "Choose meal",
-        "price": 179
-      },
-      {
-        "id": "8ebc07f9-dd0b-4929-a03f-bb25dab4b6a0",
-        "title": "Choose seats",
-        "price": 30
-      },
-      {
-        "id": "d4f4310c-a62e-4391-90bf-7910d7b84213",
-        "title": "Upgrade to comfort class",
-        "price": 132
-      },
-      {
-        "id": "34725a3e-ec85-40f8-b5ad-6803657a9759",
-        "title": "Upgrade to business class",
-        "price": 107
-      },
-      {
-        "id": "270268b1-a793-4701-8cd7-2d00d01fb604",
-        "title": "Add luggage",
-        "price": 197
-      },
-      {
-        "id": "7fabdb87-fbd8-4bf4-a6d9-34741124cc68",
-        "title": "Business lounge",
-        "price": 162
-      }
-    ]
-  },
-  {
-    "type": "check-in",
-    "offers": [
-      {
-        "id": "a29cba17-0b00-4b06-9c16-1c58d2d9ba69",
-        "title": "Choose the time of check-in",
-        "price": 60
-      },
-      {
-        "id": "3b0020db-4d7d-41e0-8d08-f2a953063c88",
-        "title": "Choose the time of check-out",
-        "price": 33
-      },
-      {
-        "id": "5e100612-fcf0-4a4e-ab19-d1432481520c",
-        "title": "Add breakfast",
-        "price": 70
-      },
-      {
-        "id": "91f9c54d-7980-49dc-a3dd-782e6fc4c522",
-        "title": "Laundry",
-        "price": 170
-      },
-      {
-        "id": "71624f8a-de35-49d6-a96b-866f3775a71a",
-        "title": "Order a meal from the restaurant",
-        "price": 60
-      }
-    ]
-  },
-  {
-    "type": "sightseeing",
-    "offers": []
-  },
-  {
-    "type": "ship",
-    "offers": [
-      {
-        "id": "f47c5269-fa5f-43d0-9ff9-701981e65bb1",
-        "title": "Choose meal",
-        "price": 124
-      },
-      {
-        "id": "fc05456e-e483-4cd0-99bf-95ca41cd24c3",
-        "title": "Choose seats",
-        "price": 171
-      },
-      {
-        "id": "7692b423-e07f-4777-aac6-aeca5bc64556",
-        "title": "Upgrade to comfort class",
-        "price": 53
-      },
-      {
-        "id": "9c2489dc-64fc-49d0-91ed-8b13cd36fd0b",
-        "title": "Upgrade to business class",
-        "price": 159
-      },
-      {
-        "id": "8e3040de-2478-474f-b61d-a8414934a824",
-        "title": "Add luggage",
-        "price": 126
-      },
-      {
-        "id": "56221ee4-0112-4fe2-a3fd-bae34d18282b",
-        "title": "Business lounge",
-        "price": 165
-      }
-    ]
-  },
-  {
-    "type": "drive",
-    "offers": [
-      {
-        "id": "dbb05b7b-44bb-49d3-99d7-5e7fc9e19bf6",
-        "title": "With automatic transmission",
-        "price": 186
-      },
-      {
-        "id": "ce623492-a21a-492b-b6d1-03b016efa4cb",
-        "title": "With air conditioning",
-        "price": 65
-      }
-    ]
-  },
-  {
-    "type": "restaurant",
-    "offers": [
-      {
-        "id": "e1f59ff7-054a-4852-bd18-72a562f4cdec",
-        "title": "Choose live music",
-        "price": 128
-      },
-      {
-        "id": "90a75790-ad6e-43f2-8fc3-cec5727b9432",
-        "title": "Choose VIP area",
-        "price": 94
-      }
-    ]
-  }
-];
